fix examples
