+++ python_modules/libraries/dagster-pandas/dagster_pandas/examples/__init__.py
@@ -1,38 +1,61 @@
-import dagstermill
 from dagster import (
     In,
-    Out,
     InputDefinition,
+    Out,
     OutputDefinition,
     config_from_files,
     file_relative_path,
     fs_io_manager,
     graph,
+    in_process_executor,
     repository,
 )
 
 from ..data_frame import DataFrame
-from .pandas_hello_world.job import pandas_hello_world
+from .pandas_hello_world.ops import always_fails_op, papermill_pandas_hello_world, sum_op, sum_sq_op
 
 
-def nb_test_path(name):
-    return file_relative_path(__file__, "notebooks/{name}.ipynb".format(name=name))
+@graph
+def pandas_hello_world_fails():
+    always_fails_op(sum_sq_op=sum_sq_op(sum_df=sum_op()))
 
 
-hello_world = dagstermill.define_dagstermill_solid(
-    name="papermill_pandas_hello_world",
-    notebook_path=nb_test_path("papermill_pandas_hello_world"),
-    input_defs=[InputDefinition(name="df", dagster_type=DataFrame)],
-    output_defs=[OutputDefinition(DataFrame)],
+pandas_hello_world_fails_test = pandas_hello_world_fails.to_job(executor_def=in_process_executor)
+
+
+@graph
+def pandas_hello_world():
+    sum_sq_op(sum_op())
+
+
+pandas_hello_world_test = pandas_hello_world.to_job(
+    config=config_from_files(
+        [
+            file_relative_path(
+                __file__, "pandas_hello_world/environments/pandas_hello_world_test.yaml"
+            )
+        ]
+    ),
+    executor_def=in_process_executor,
+)
+
+pandas_hello_world_prod = pandas_hello_world.to_job(
+    config=config_from_files(
+        [
+            file_relative_path(
+                __file__, "pandas_hello_world/environments/pandas_hello_world_prod.yaml"
+            )
+        ]
+    )
 )
 
 
 @graph
-def papermill_pandas_hello_world():
-    hello_world()
+def papermill_pandas_hello_world_graph():
+    papermill_pandas_hello_world()
 
 
-papermill_pandas_hello_world_test = papermill_pandas_hello_world.to_job(
+papermill_pandas_hello_world_test = papermill_pandas_hello_world_graph.to_job(
     resource_defs={"io_manager": fs_io_manager},
     config=config_from_files(
         [
@@ -44,7 +67,7 @@ papermill_pandas_hello_world_test = papermill_pandas_hello_world.to_job(
     ),
 )
 
-papermill_pandas_hello_world_prod = papermill_pandas_hello_world.to_job(
+papermill_pandas_hello_world_prod = papermill_pandas_hello_world_graph.to_job(
     resource_defs={"io_manager": fs_io_manager},
     config=config_from_files(
         [
@@ -58,5 +81,10 @@ papermill_pandas_hello_world_prod = papermill_pandas_hello_world.to_job(
 
 
 @repository
-def test_dagstermill_pandas():
-    return [papermill_pandas_hello_world_prod, pandas_hello_world]
+def dagstermill_pandas_test_repo():
+    return [papermill_pandas_hello_world_test, pandas_hello_world_test]
+
+
+@repository
+def dagstermill_pandas_prod_repo():
+    return [papermill_pandas_hello_world_prod, pandas_hello_world_prod]
