+++ python_modules/libraries/dagster-pandas/dagster_pandas/examples/workspace.yaml
@@ -1,4 +1,3 @@
 load_from:
   - python_module:
       module_name: dagster_pandas.examples
-      attribute: define_pandas_repository
