+++ python_modules/libraries/dagster-pandas/dagster_pandas_tests/test_config_driven_df.py
@@ -2,13 +2,7 @@ import os
 
 import pandas as pd
 import pytest
-from dagster import (
-    DagsterInvalidConfigError,
-    In,
-    Out,
-    graph,
-    op,
-)
+from dagster import DagsterInvalidConfigError, In, Out, graph, op
 from dagster.utils import file_relative_path
 from dagster.utils.test import get_temp_file_name
 from dagster_pandas import DataFrame
