+++ python_modules/libraries/dagster-pandas/dagster_pandas_tests/test_data_frame.py
@@ -9,7 +9,6 @@ from dagster import (
     Out,
     Output,
     Selector,
-    build_op_context,
     check_dagster_type,
     dagster_type_loader,
     dagster_type_materializer,
