+++ python_modules/libraries/dagster-pandas/dagster_pandas_tests/test_pandas_metadata.py
@@ -1,5 +1,4 @@
 import pandas as pd
-
 from dagster import DagsterEventType, In, file_relative_path, graph, op
 from dagster_pandas import DataFrame
