+++ python_modules/libraries/dagster-pandas/dagster_pandas_tests/test_pandas_user_error.py
@@ -3,13 +3,7 @@
 import dagster_pandas as dagster_pd
 import pandas as pd
 import pytest
-from dagster import (
-    DagsterTypeCheckDidNotPass,
-    In,
-    Out,
-    graph,
-    op,
-)
+from dagster import DagsterTypeCheckDidNotPass, In, Out, graph, op
 
 
 def test_wrong_output_value():
