+++ python_modules/libraries/dagster-pandas/dagster_pandas_tests/test_structured_df_types.py
@@ -1,10 +1,4 @@
-from dagster import (
-    DagsterType,
-    Out,
-    Output,
-    graph,
-    op,
-)
+from dagster import DagsterType, Out, Output, graph, op
 from dagster_pandas.constraints import (
     ColumnWithMetadataException,
     ConstraintWithMetadataException,
